requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
web3==6.20.0
hiero-sdk-python==1.3.0
python-dotenv==1.0.0
requests==2.32.0
groq==0.9.0
pydantic==2.9.0